requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
async-timeout==5.0.1
attrs==25.4.0
audioread==3.0.1
av==15.1.0
babel==2.17.0
bangla==0.0.5
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
build==1.3.0
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.11
Cython==3.1.4
dateparser==1.1.8
decorator==5.2.1
deepspeed==0.18.0
diffusers==0.35.1
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
g2pkk==0.1.2
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.75.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
hangul-romanize==0.1.0
hf-xet==1.1.10
hjson==3.1.0
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.2
jsonlines==1.2.0
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.45.1
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
murmurhash==1.0.13
networkx==2.8.8
ninja==1.13.0
nltk==3.9.2
num2words==0.5.14
numba==0.62.1
numpy==1.22.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==1.5.3
peft==0.17.1
pillow==11.3.0
pip-tools==7.5.1
platformdirs==4.4.0
pooch==1.8.2
preshed==3.0.10
propcache==0.4.0
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.5
pypinyin==0.55.0
pyproject_hooks==1.2.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.11.4
sentencepiece==0.2.1
sentry-sdk==2.40.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
SudachiDict-core==20250825
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
timm==1.0.20
tokenizers==0.21.4
tomli==2.3.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
trainer==0.0.36
transformers==4.48.0
triton==3.4.0
TTS==0.22.0
typeguard==4.4.4
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
urllib3==2.5.0
wandb==0.22.2
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
lpips==0.1.4
torchmetrics==1.8.2
torch-fidelity==0.3.0